modified:   src/animation.c 	modified:   src/animation.h 	modified:   src/assets.c 	modified:   src/assets.h 	modified:   src/behaviour.c 	modified:   src/behaviour.h 	modified:   src/box_collision.c 	modified:   src/box_collision.h 	modified:   src/circle_collision.c 	modified:   src/circle_collision.h 	modified:   src/collision.c 	modified:   src/collision.h 	modified:   src/content.c 	modified:   src/content.h 	modified:   src/entity.c 	modified:   src/entity.h 	modified:   src/event.c 	modified:   src/event.h 	modified:   src/kinematics.c 	modified:   src/kinematics.h 	modified:   src/sdlutils.c 	modified:   src/sdlutils.h 	modified:   src/shmup.c 	deleted:    src/shot_patterns.c 	deleted:    src/shot_patterns.h 	modified:   src/stage.h

+++ src/event.c
@@ -43,7 +43,7 @@ raiseEvent(const EventManager *mgr, const void *args)
         printf("EVENT RAISED: %s\n", mgr->label);
     else
         puts("ANONYMOUS EVENT RAISED");
-    for (int i = 0; i < mgr->subscribers->count; i++) {
+    for (size_t i = 0; i < mgr->subscribers->count; i++) {
         sub = (Subscriber *)getElementArray(mgr->subscribers, i);
         printf("Notifying: %s\n", sub->label == NULL ? "ANONYMOUS" : sub->label);
         sub->eventFunc(mgr->parent, sub->subscriber, args);
